requirements.txt
# Core
streamlit==1.24.0
pandas==2.0.3
numpy==1.24.3
plotly==5.15.0
scikit-learn==1.3.0

# Data Processing
openpyxl==3.1.2
python-dotenv==1.0.0
pymongo==4.4.0
sqlalchemy==2.0.17
psycopg2-binary==2.9.6
redis==4.6.0
elasticsearch==8.8.0

# Visualization
seaborn==0.12.2
matplotlib==3.7.1

# Machine Learning
prophet==1.1.4
statsmodels==0.14.0

# NLP & AI
transformers==4.30.2
torch==2.0.1
googletrans==3.1.0a0

# Security
cryptography==41.0.1
python-jwt==4.0.0
bcrypt==4.0.1

# Report Generation
jinja2==3.1.2
pdfkit==1.0.0
weasyprint==59.0

# Cache & Performance
joblib==1.3.1
cachetools==5.3.1

# API & Web
requests==2.31.0
websockets==11.0.3
httpx==0.24.1
fastapi==0.100.0
uvicorn==0.22.0

# Development
pytest==7.4.0
black==23.3.0
mypy==1.4.1
pylint==2.17.4